pyproject.toml
[project]
name = "kltn-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.5.2",
    "fastapi[standard]>=0.115.12",
    "faststream[rabbit]>=0.5.40",
    "logfire[sqlalchemy]>=3.15.0",
    "minio>=7.2.15",
    "psycopg[binary,pool]>=3.2.7",
    "pydantic-ai>=0.1.10",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.40",
    "structlog>=25.3.0",
]
